mc_backend/requirements.txt
alembic==1.4.2
aniso8601==8.0.0
appdirs==1.4.4
attrs==19.3.0
beautifulsoup4==4.9.0
black==19.10b0
blinker==1.4
boto==2.49.0
boto3==1.12.49
botocore==1.15.49
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
clickclick==1.2.2
conllu==2.3.2
connexion==2.7.0
coverage==5.1
cycler==0.10.0
decorator==4.4.2
docutils==0.15.2
Flask==1.1.2
Flask-Cors==3.0.8
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.1
frozendict==1.2
future==0.18.2
gensim==3.8.2
graphannis==0.27.0
gunicorn==20.0.4
html5lib==1.0.1
idna==2.9
inflection==0.4.0
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.9.5
jsonschema==3.2.0
kiwisolver==1.2.0
LinkHeader==0.4.3
lxml==4.5.0
Mako==1.1.2
MarkupSafe==1.1.1
matplotlib==3.2.1
MyCapytain==3.0.2
networkx==2.4
numpy==1.18.3
OpenAlchemy==1.1.0
openapi-spec-validator==0.2.8
pathspec==0.8.0
Pillow==7.1.2
prance==0.18.2
psycopg2==2.8.5
psycopg2-binary==2.8.5
pycparser==2.20
PyLD==2.0.2
pyparsing==2.4.7
PyPDF2==1.26.0
pyrsistent==0.16.0
python-dateutil==2.8.1
python-docx==0.8.10
python-dotenv==0.13.0
python-editor==1.0.4
python-rapidjson==0.9.1
pytz==2020.1
PyYAML==5.3.1
rdflib==5.0.0
rdflib-jsonld==0.5.0
regex==2020.5.7
reportlab==3.5.42
requests==2.23.0
s3transfer==0.3.3
scipy==1.4.1
semver==2.10.1
six==1.14.0
smart-open==2.0.0
soupsieve==2.0
SQLAlchemy==1.3.16
swagger-ui-bundle==0.0.6
toml==0.10.0
typed-ast==1.4.1
typing==3.7.4.1
urllib3==1.25.9
webencodings==0.5.1
Werkzeug==1.0.1
xhtml2pdf==0.2.4
